Updated the calendar functionality
Updated the calendar to add css based on date selected. Also cleaned up
the template to fix the a tag to close correctly around the the div.day
. In addition to those changes I removed the internal link that was
causing the date and added the to the overall div.day class to clean up
the code and streamline the calendar output.