Add logic to stop conflict with MyLCCCInfoFeed
Both this plugin and MyLCCCInfoFeed have these two files

plugin_functions.php
rest-api-fetch.php

in common. So to stop them from causing a fatal error when LCCC News is
activated I added logic to test to see if MyLCCCInfoFeed is active in on
the site or on the network. If it is then LCCC News will not require
these two files. If it is not the files are then loaded by LCCC News